WIP: Vendor circom-compat without wasmer
