+++ crates/ark-circom/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ark-circom"
+name = "semaphore-rs-ark-circom"
 version.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -17,8 +17,16 @@ ark-serialize.workspace = true
 ark-std.workspace = true
 ark-bn254.workspace = true
 ark-groth16.workspace = true
+ark-poly.workspace = true
 ark-crypto-primitives.workspace = true
 
+ruint.workspace = true
+serde_json.workspace = true
 byteorder.workspace = true
 num-traits.workspace = true
 num-bigint.workspace = true
+
+[features]
+# This features does nothing but the ark_std::cfg_into_iter expectes it
+parallel = []
+
