+++ crates/ark-zkey/Cargo.toml
@@ -11,13 +11,14 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
+semaphore-rs-ark-circom.workspace = true
+
 color-eyre.workspace = true
 memmap2.workspace = true
 
 ark-serialize.workspace = true
 ark-bn254.workspace = true
 ark-groth16.workspace = true
-ark-circom.workspace = true
 ark-relations.workspace = true
 ark-ff.workspace = true
 ark-ec.workspace = true
