+++ crates/semaphore/Cargo.toml
@@ -14,6 +14,7 @@ categories.workspace = true
 # Internal
 semaphore-rs-utils.workspace = true
 semaphore-rs-ark-zkey.workspace = true
+semaphore-rs-ark-circom.workspace = true
 semaphore-rs-proof = { workspace = true, features = ["ark"] }
 semaphore-rs-poseidon.workspace = true
 semaphore-rs-hasher.workspace = true
@@ -47,7 +48,6 @@ zeroize.workspace = true
 
 # Ark
 ark-bn254.workspace = true
-ark-circom.workspace = true
 ark-ec.workspace = true
 ark-ff.workspace = true
 ark-groth16.workspace = true
