+++ crates/trees/Cargo.toml
@@ -14,6 +14,7 @@ categories.workspace = true
 # Internal
 semaphore-rs-hasher.workspace = true
 semaphore-rs-storage.workspace = true
+semaphore-rs-ark-circom.workspace = true
 
 # 3rd Party
 bytemuck.workspace = true
@@ -32,7 +33,6 @@ tiny-keccak = { workspace = true, features = ["sha3"] }
 
 # Ark
 ark-bn254.workspace = true
-ark-circom.workspace = true
 ark-ec.workspace = true
 ark-ff.workspace = true
 ark-groth16.workspace = true
